Fix for parsing metadata with duplicate labels
CBOR doesn't strictly restrict duplicate keys unless strict mode is
enabled so this allowed such data to be allowed on-chain

this fixes #73

we don't keep the duplicate metadatum consistent with a previous fix
with duplicate keys in metadata maps. how to handle this isn't
well-defined by the specs and seems to be  an oversight/bug in the node
code to not reject it.